+++ ydb/core/tablet_flat/ut/ya.make
@@ -50,7 +50,6 @@ SRCS(
     ut_screen.cpp
     ut_bloom.cpp
     ut_shared_sausagecache.cpp
-    ut_shared_sausagecache_memtable.cpp
     ut_slice.cpp
     ut_slice_loader.cpp
     ut_versions.cpp
